+++ core/mod.ts
@@ -7,6 +7,7 @@ export * from "./tokens.ts";
 export * from "./migrator.ts";
 export * from "./postgres.ts";
 export * from "./random.ts";
+export * from "./server-sent-events.ts";
 export * from "./store.ts";
 export * from "./structures.ts";
 export * from "./terminator.ts";